requirements.txt
Flask==3.0.0
google-generativeai==0.3.1
Pillow==10.1.0
python-dotenv==1.0.0
cryptography==41.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
